The sports/1, i.e. when hitting 'soccer' it pulls from the soccer API, also renamed 'users.js' 'sports.js' file in 'Routes folder' users_controller.js and sports_controller.js, AND changed usersRoute to sportsRoute paths respectively.
